api/requirements.txt
Flask==0.12.2
flask-restful==0.3.6
ptvsd==4.3.2
mysql-connector==2.2.9